requirements.txt
brazilnum==0.8.8
Django==2.0.4
django-model-utils==3.1.1
djangorestframework==3.8.2
pytz==2018.4
